+++ src/tokens/mod.rs
@@ -1,11 +1,9 @@
 mod constants;
-mod default_token_action;
-mod determine_token_action;
 pub mod dm_token;
-mod string_in_quote;
-mod string_special_escape;
+mod get_token_action;
+mod parse_character;
+mod start_tokenize;
 mod token_action;
-mod token_handling;
 
 #[cfg(test)]
 mod tests;
