+++ src/tokens/tests/default_token_action.rs
@@ -2,7 +2,7 @@
 mod tests {
     use crate::tokens::{
         constants::{SYM_MATH, SYM_OTHR, SYM_QUTE},
-        default_token_action::get_default_token_action,
+        parse_character::defaults::handle_defaults,
         token_action::TokenAction,
     };
 
@@ -13,10 +13,9 @@ mod tests {
         symbols.extend(SYM_OTHR);
 
         for symbol in symbols {
-            assert!(get_default_token_action(symbol, "text") == TokenAction::StartNewToken);
+            assert!(handle_defaults(symbol, "text") == TokenAction::StartNewToken);
             assert!(
-                get_default_token_action(symbol, symbol.to_string().as_str())
-                    == TokenAction::ContinueToken
+                handle_defaults(symbol, symbol.to_string().as_str()) == TokenAction::ContinueToken
             );
         }
     }
@@ -24,41 +23,40 @@ mod tests {
     #[test]
     fn test_quote_symbols() {
         for symbol in SYM_QUTE {
-            assert!(get_default_token_action(*symbol, "text") == TokenAction::IsolateToken);
+            assert!(handle_defaults(*symbol, "text") == TokenAction::IsolateToken);
             assert!(
-                get_default_token_action(*symbol, symbol.to_string().as_str())
-                    == TokenAction::IsolateToken
+                handle_defaults(*symbol, symbol.to_string().as_str()) == TokenAction::IsolateToken
             );
         }
     }
 
     #[test]
     fn test_whitespace_transition() {
-        assert!(get_default_token_action(' ', "text") == TokenAction::StartNewToken);
-        assert!(get_default_token_action(' ', " ") == TokenAction::ContinueToken);
-        assert!(get_default_token_action('\t', "") == TokenAction::ContinueToken);
-        assert!(get_default_token_action(' ', "\t") == TokenAction::StartNewToken);
+        assert!(handle_defaults(' ', "text") == TokenAction::StartNewToken);
+        assert!(handle_defaults(' ', " ") == TokenAction::ContinueToken);
+        assert!(handle_defaults('\t', "") == TokenAction::ContinueToken);
+        assert!(handle_defaults(' ', "\t") == TokenAction::StartNewToken);
     }
 
     #[test]
     fn test_digit_transition() {
-        assert!(get_default_token_action('1', "text") == TokenAction::ContinueToken);
-        assert!(get_default_token_action('2', "123") == TokenAction::ContinueToken);
-        assert!(get_default_token_action('a', "123") == TokenAction::ContinueToken);
+        assert!(handle_defaults('1', "text") == TokenAction::ContinueToken);
+        assert!(handle_defaults('2', "123") == TokenAction::ContinueToken);
+        assert!(handle_defaults('a', "123") == TokenAction::ContinueToken);
     }
 
     #[test]
     fn test_no_transition() {
-        assert!(get_default_token_action('a', "text") == TokenAction::ContinueToken);
-        assert!(get_default_token_action('3', "12") == TokenAction::ContinueToken);
-        assert!(get_default_token_action('-', "-") == TokenAction::ContinueToken);
+        assert!(handle_defaults('a', "text") == TokenAction::ContinueToken);
+        assert!(handle_defaults('3', "12") == TokenAction::ContinueToken);
+        assert!(handle_defaults('-', "-") == TokenAction::ContinueToken);
     }
 
     #[test]
     fn test_underscores_in_alphanumeric() {
-        assert!(get_default_token_action('_', "text") == TokenAction::ContinueToken);
-        assert!(get_default_token_action('_', "text1") == TokenAction::ContinueToken);
-        assert!(get_default_token_action('_', "text2_") == TokenAction::ContinueToken);
-        assert!(get_default_token_action('_', "$") == TokenAction::StartNewToken);
+        assert!(handle_defaults('_', "text") == TokenAction::ContinueToken);
+        assert!(handle_defaults('_', "text1") == TokenAction::ContinueToken);
+        assert!(handle_defaults('_', "text2_") == TokenAction::ContinueToken);
+        assert!(handle_defaults('_', "$") == TokenAction::StartNewToken);
     }
 }
